Move strata_link_to into Strata::LinksHelper
strata_link_to has a passthrough default that produces plain links —
it doesn't belong in a module named “Buttons.” Splits the helper into
Strata::LinksHelper (just strata_link_to) and Strata::ButtonsHelper
(just strata_button_to), matching the engine's existing
one-topic-per-helper pattern. Both are included via
Strata::ApplicationHelper so they remain auto-available in views.
